refactor: add product validation and transformation methods on product model

+++ src/controllers/admin.product.controller.ts
@@ -1,51 +1,22 @@
-import { validateProduct, validateUpdateProduct } from "../lib/validateProduct";
 import { catchAsyncError } from "../middlewares/catchAsyncError";
 import Product from "../models/product.model";
 import { ErrorHandler } from "../lib/errorHandler";
 import { deleteImage, uploadImage } from "../lib/cloudinary";
-import {
-  ProductValidationBody,
-  UpdateProductValidationBody,
-} from "../lib/productValidationSchemas";
+import { CreateProductBody, UpdateProductBody } from "../types/product";
+import { isNumberArray, isStringArray } from "../lib/validators";
 
 export const createProduct = catchAsyncError<
   unknown,
   unknown,
-  ProductValidationBody
+  CreateProductBody
 >(async (req, res) => {
-  validateProduct(req.body);
-
   const { images, ...productDetails } = req.body;
   let product = new Product({
     ...productDetails,
     owner: req.user._id,
   });
 
-  product.features.splice(10);
-  product.tags.splice(5);
-  switch (product.category) {
-    case "mobile":
-      break;
-    case "laptop":
-      break;
-    case "electronics":
-      break;
-    default:
-      delete product["ram"];
-  }
-
-  product.sizes = product.sizes.filter((size) => {
-    return (
-      size === "sm" ||
-      size === "md" ||
-      size === "lg" ||
-      size === "xl" ||
-      size === "2xl"
-    );
-  }) as typeof product.sizes;
-  product.colors.splice(7);
-
-  if (images) {
+  if (images && isStringArray(images)) {
     for (const image of images.slice(0, 5)) {
       const res = await uploadImage(image);
       if (res)
@@ -66,13 +37,10 @@ export const createProduct = catchAsyncError<
 export const updateProduct = catchAsyncError<
   { id: string },
   unknown,
-  UpdateProductValidationBody
+  UpdateProductBody
 >(async (req, res, next) => {
   const product = await Product.findById(req.params.id);
-  if (!product)
-    return next(new ErrorHandler("Product with this id doesn't exist", 400));
-
-  validateUpdateProduct(req.body);
+  if (!product) return next(new ErrorHandler("Product doesn't exist", 400));
 
   const { images, ...productDetails } = req.body;
   for (const key of Object.keys(productDetails)) {
@@ -80,7 +48,13 @@ export const updateProduct = catchAsyncError<
     product[key] = productDetails[key];
   }
 
-  if (images) {
+  if (
+    images &&
+    images.add &&
+    images.indexesToDelete &&
+    isStringArray(images.add) &&
+    isNumberArray(images.add)
+  ) {
     let delIndex = 0;
     for (const image of images.add.slice(0, 5)) {
       await deleteImage(
